backend/requirements.txt
Django==5.2.1
djangorestframework==3.15.2
djoser==2.2.3
drf-extra-fields==3.2.1
webcolors==1.11.1
django-cors-headers==3.13.0
Pillow==10.4.0
django-bootstrap5==22.2
djangorestframework-simplejwt==5.3.1
pytest-pythonpath==0.7.3
pytest-django==4.4.0
django-filter==24.3
psycopg2-binary==2.9.9
gunicorn==23.0.0
python-dotenv==1.0.1
shortuuid==1.0.11
drf-yasg==1.21
reportlab==4.2.5
